readbuffer: add read timeout
Read() may wait forever if no deadline is set and no calls to Close()
are made, causing goroutine and memory leaks.

This patch adds a read timeout in case no deadline is set, calling
Close() if more than DefaultReadTimeout elapses. It is meant to be
a last line of defense if all else fails closing the connection.

Such a problematic condition was observed in Rancher 2.7.6, specifically
with remotedialer.pipe() calls blocked by Read() on the rancher pods.
In that scenario, Close() is expected to be called upon receiving
Error messages, but those might never be delivered at all, eg. because
of werite timeouts at the sending end.

Signed-off-by: Silvio Moioli <silvio@moioli.net>